requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
decorator==5.1.1
Django==4.0.1
djangorestframework==3.13.1
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
nominatim==0.1
openrouteservice==2.3.3
pytz==2021.3
ratelim==0.1.6
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
